requirements.txt
FastAPI>=0.95.0
python-osc>=1.7.4
python-dotenv>=0.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.20.0
requests>=2.28.0
pytest>=7.0.0